Remove shared_ptr_helper
- The logic encapsulated in "shared_ptr_helper" was a trivial, i.e. just
  a oneliner. Removed it as per discussion in the PR.

- Replaced the current machinery for shared_ptr_helper (inheritance +
  friend declaration) by a direct implementation of "create()" in the
  user classes, this time based on make_shared<>() instead of
  shared_ptr<>() to save one heap allocation.

- To disallow bypassing T::create(), i.e. direct instantiation of class
  T, the constructors of T were in almost all cases declared private or
  protected. However, unlike shared_ptr<T>(), make_shared<T>() needs
  access to a public constructor of T. I therefore applied the passkey
  idiom which adds a public forwarding constructor that can only be
  called with a 'passkey' private to T. ( https://stackoverflow.com/a/8147326).
  As a result, we get exactly the same behavior as before.

- There were a handful of classes inheriting from shared_ptr_helper
  which had a public constructor, e.g. StorageMySQL. I assume this was
  unintentional as it allows direct construction. I made these
  constructors private too.

- Also moved the shared_ptr traits to the only place where they are
  actually needed (typeid_cast.h) to reduce include bloat.

+++ src/Storages/tests/gtest_storage_log.cpp
@@ -36,7 +36,7 @@ DB::StoragePtr createStorage(DB::DiskPtr & disk)
 
     StoragePtr table = StorageLog::create(
         "Log", disk, "table/", StorageID("test", "test"), ColumnsDescription{names_and_types},
-        ConstraintsDescription{}, String{}, false, 1048576);
+        ConstraintsDescription{}, String{}, false, static_cast<size_t>(1048576));
 
     table->startup();
